// src/interfaces/api.types.generated.d.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Scalar for receiving email from client */
  Email: any;
};
// ...
export type CarFeatureType = {
  __typename?: 'CarFeatureType';
  id: Scalars['String'];
  name: Scalars['String'];
  type: CarFeatureTypeEnum;
};

export enum CarFeatureTypeEnum {
  Extra = 'Extra',
  Other = 'Other'
}
// ...
export enum CarMotorEnum {
  Electric = 'Electric',
  Hybrid = 'Hybrid'
}
// ...
export enum CarTypeEnum {
  Car = 'Car',
  Van = 'Van'
}

export enum CarValueEnum {
  Between10And20 = 'Between10And20',
  Between20And30 = 'Between20And30',
  Between30And40 = 'Between30And40',
  Between40And60 = 'Between40And60',
  Between60And75 = 'Between60And75',
  MoreThan75 = 'MoreThan75'
}